chore(storybook): move token loader into own file
